fix(dashboard): align view types with protocol
